refactor: ProductOptionList 컴포넌트 분리
